+++ changelog.md
@@ -1,3 +1,15 @@
+# v0.0.41
+- More features are available to load, added gamemode and groundLayer
+
+# v0.0.4
+- Experimental loading of saved progress has been added to the game. The mod only loads the position of the player.
+
+# v0.0.31
+- The game doesn't crash now :3
+- 
+# v0.0.3
+- Saving progress is now fully functional. Restoring the progress will not be present in this update, or be extremely experimental
+
 # v0.0.2
 - Added a button in the pause menu that shows current save status, and broadcasting progress to file.